requirements.txt
flask>=2.0,<3.0
flask-socketio>=5.0,<7.0
eventlet>=0.33.0
RPi.GPIO>=0.7.0
opencv-python
